tracker/requirements.txt
httpx==0.27.0
psutil==5.9.8
pystray==0.19.5
pywin32==306
Pillow==10.3.0
